Provide hex/bytes utils as part of js-waku
